refactor: split the service into components
closes #145
