+++ Dockerfile
@@ -1,19 +1,3 @@
-FROM node:8.11.1-alpine as react-builder
-
-# Uncomment the following line to allow live updating of files for development.
-# ADD . /app
-WORKDIR /app
-
-COPY src/ui/package.json src/ui/package-lock.json /app/
-
-RUN npm install --silent
-
-COPY src/ui/public /app/public
-COPY src/ui/src /app/src/
-
-RUN npm run-script build
-
-
 FROM python:3.7-alpine
 
 LABEL maintainer="info@datascience.ch"
@@ -25,21 +9,18 @@ RUN apk update && \
     addgroup -g 1000 kyaku && \
     adduser -S -u 1000 -G kyaku kyaku
 
-# Copy static assets from react-builder stage
-COPY --from=react-builder /app/build /app/static
-
 # Install all packages
-COPY Pipfile Pipfile.lock /app/
-WORKDIR /app
+COPY Pipfile Pipfile.lock /renku-notebooks/
+WORKDIR /renku-notebooks
 RUN pipenv install --system --deploy
 
-# Move the service source code
-COPY src /app/src
-ENV FLASK_APP=/app/src/notebooks_service.py
+COPY renku_notebooks /renku-notebooks/renku_notebooks
+# Set up the flask app
+ENV FLASK_APP=/renku-notebooks/renku-notebooks:create_app
 
 # Switch to unpriviledged user
 USER kyaku
 
-CMD ["gunicorn", "-b 0.0.0.0:8000", "src:app", "-k gevent"]
+CMD ["gunicorn", "-b 0.0.0.0:8000", "renku_notebooks.wsgi:app", "-k gevent"]
 
 HEALTHCHECK --interval=20s --timeout=10s --retries=5 CMD curl -f http://localhost:8000/health
