+++ renku_notebooks/api/__init__.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2018 - Swiss Data Science Center (SDSC)
+# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Eidgenössische Technische Hochschule Zürich (ETHZ).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+""" Notebooks service APIs """
+import os
+from urllib.parse import urljoin
+
+from flask import Blueprint, Response
+
+from . import auth, notebooks
+from .. import config
+
+bp = Blueprint('notebooks_service', __name__)
+
+
+@bp.route('/health')
+def health():
+    """Just a health check path."""
+    return Response('service running under {}'.format(config.SERVICE_PREFIX))
+
+
+blueprints = [bp, auth.bp, notebooks.bp]
