+++ renku_notebooks/wsgi.py
@@ -1,0 +1,3 @@
+from renku_notebooks import create_app
+
+app = create_app()
