--- src/__init__.py
@@ -1,3 +1,0 @@
-from . import notebooks_service
-
-app = notebooks_service.app
