--- src/templates/server_status.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<!-- Copyright 2018 - Swiss Data Science Center (SDSC)
-A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
-Eidgenössische Technische Hochschule Zürich (ETHZ).
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. -->
-<head>
-  <title>server {{server_name}} status</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css">
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.0/umd/popper.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js"></script>
-</head>
-<body>
-
-<div class="jumbotron text-center">
-  <h1>Server for {{namespace}}/{{project}}/{{commit_sha}} status: {{status}}</h1>
-  {% if status == "not found" %}
-    <h3>You can try to launch it again from <a style="color: #5561a6" href="/">Renku</a>.</h3>
-    <h3>If the launch is underway, you will soon be automatically redirected.</h3>
-    <meta http-equiv="refresh" content="2">
-  {% elif status == "waiting" or status == "spawning" %}
-    <h3>🚀 Your server is almost ready! 🚀</h3>
-    <meta http-equiv="refresh" content="2">
-  {% elif status == "stopping" %}
-    <h3>The server is in the process of terminating.</h3>
-    <meta http-equiv="refresh" content="2">
-  {% endif %}
-</div>
-
-</body>
-</html>
