--- src/ui/README.md
@@ -1,6 +1,0 @@
-# Renku Notebooks UI
-
-A minimal UI for administering the service that provides interactive notebooks for
-the Renku platform.
-
-This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
